♻️ rename `types` to `models`
